requirements.txt
fastapi==0.118.0
 uvicorn[standard]==0.24.0
 pydantic==2.5.0
 tortoise-orm[asyncpg]==0.25.1
